Attempt to parse e:e=e the same as e:e:=e.
This uses a copy of Control.Monad.Combinators.Expr that allows
the operator parser to peek at the first operand.
This way the usual '=' in 'e:e=e' will fail to parse in case
the first operand is 'e:e', and a different clause with the
same precedence as ':=' kicks in.