Enum with iteration, comparison, and coerce
Parsing is somewhat fake, currently only verified that : and indentation with one value per line works.

enum:
  A
  B

Is parsed in the same way as module.

Enum (List [Name "A", Name "B"])

Desugare then uses fixEnum' to verify that the list only contains names, and then translates it to look something line the following (in pseudo verse)

module:
  A := enumValue 0
  B := enumValue 1

The representation is

Enum enumLabel hashMap [ArchetypeName (Name "A") :=: IfArchetypeName ... (EnumValue enumLabel 0),
                        ... ]
Eval evaluates this to:

Enum enumLabel hashMap [A, ...] [the value of A, ...]

The two lists are for iteration and pretty printing. An alternative implementation would be to extract and sort the enum values from the hash map.