The parser is now organized more like TimGrammar.
One of the exeption is comparisons that are parsed as a list of operators and operands. This list is then processed to produce either a comparison expression, something with attributes, or a parse error.
The parse error is not very nice but useful to se what it get stuck on.

The verseparser can now take paths as arguments.
If the path ends in ".verse" or ".versetest" then it's parsed.
If the path is a directory, then verseparse recurse into the directory.
Non-directory paths not ending in ".verse" or ".versepath" are silently ignored.

Fix for missing undents due to missing newline at end of verse-file.

Changed some test files since "X|:Y" wasn't accepted by TimGrammar.

There are both a '.' and a '. ' tokens in TimGrammar. Added the latter but it currently doesn't handle beeing followed by a comment.

Add possiblity for attributes on 'enum' and 'var', and in many other places.

A lot of the reserved words are removed from the lexer. They are treated as macros in the parser, and then recognized in Rewrite. This is done to reduce the amount of code in the parser. Might be possible to clean this up later.

Known bug in lexer with identifiers that start with a reserved word!?

Parse all.verse and can be used to run all tests. Parse a lot of the versetests files in Test, but still fail on many. Need support for @, and there are still spurious errors.