Parsec implementation of TimGrammar in Parse2.hs
The parser is in a new file since it doesn't share any history with the existing parser. This also makes it easier to compare output while debugging problems with the interaction between the new parser and the reference implementation.
The cabal target verseparse can run with either parser (--old or --new), and either only parse or also run the rewriter (--parse or --rewrite). If --verbose is used then the output is pretty printed, --show prints using Show.

Parse2 is written to follow the grammar as much as possible, but markup is not implemented yet. Other indentional differences due to existing code are:
  * Expr allows ';' after '@'-attributes
  * Block indentation does not end due to lines with only space and tab characters

Known differences beween the old parser and the new are:
  * Less indented comments breaks indentation
  * The use of prefix ':' now follows the grammar, add parantheses to the code if old code fail.
  * Error messages are different, Parsec does not report ranges, only positions.

Parser2 has more backtracking than is healthy, but removing them will take the parser further from the grammar.
I still think we need to do this in a future fix.