Add truth{e} as a primtive.
truth{e} acts like a map (i.e. a function) mapping e to e.
It would be possible to encode this as '\x.x=e; e' but
then unification of truths would not work.
So instead it is a primitive, behaving much like a singleton array.

The is no verification rule for truth yet.