repl: add repl variables
This is a patch after landing #72 we made the decision to modify the output
store to be a variable store. Why? Because the languages the repl call do not
yet have any kind of local state. So adding such state to the repl is a stop gap
while we work on the denotational semantics.