tester: rename timDensem$ -> tim$, densem funs can be marked
In 9a3169c0b0ae55ac4f9042b3a908141eb8c29e52 I landed semantic functions that
could be tested in the tester but the names are all a bit too cumbersome. So
this commit renames them as follows:

- timDensem$ -> tim$
- slsDensem$ -> sls$
- elsDensem$ -> els$
- dlsDensem$ -> dls$

in addition to that, the implementation in
9a3169c0b0ae55ac4f9042b3a908141eb8c29e52 could not mark testds cases as `pass`,
`fail`, `skip` or `broken`. This commit fixes that by not abusing `TestType` to
add the tag for the right densem function. In this mannner, this version of the
implementation is more orthogonal to the rest of the tester.