tester: remove suffixed '$' in densem functions
In 3169a055756367976df2a6d6ee0f991c9fe144fe I renamed these to shorter names but
kept the suffixed '$' to indicate that these functions are built into the tester
and are not defined in the .versetest file.

However, Lennart pointed out that '$' is not a valid symbol in an identifier
according to the Verse Spec 0.15 and therefore any .versetest file that uses a
densem function is no longer a valid verse program!

This commit repairs that property by removing the suffixed '$'