parser: tester: Parser supports basic splice patterns
- in pursuit of #86
- in pursuit of #66

This commit adds tuple and splice patterns to the parser. See the above issues
for context. These patterns occur in our tests and spec but were unimplemented
in the parser. There is still a lot of cleanup to do in the rewriter. I've
marked these with TODOs. Additionally, this commit removes a lot of variable
shadowing in the rewriter and the parser. With these additions we are down to 66
skipped testes out of 214