parser: lib parser implements check<effect>
Allows tests `check[1-9]` to pass in the test.

In pursuit of #66
In pursuit of #95

Notice that we have to catch the root `Inst` rather than implementing `ExpSpecs`
directly and using recursion. While unfortunate this is purposeful. The
recursive version loses the body of the check because that is the right child of
`Inst` and therefore the sister node of the `ExpSpecs`. So in the recursive
case, when processing the `ExpSpecs` we don't have access to the sister node. So
by directly matching on this case we can craft a proper `Check Effect Body` and
have the `Body` be the recursive case instead of using some kind of sentinal
(like List [] or something) thats later dropped for the body.