parser: add string literals, Splice6, tester tests pass with parser
closes #95, #66

This commit adds string literals without string interpolation to the rewriter
language. This then is transformed into a FrontEnd.Expr string literal.

The previous behavior was intepreting string literals as tuples of
characters. This is a workable solution and would give location data for each
character read by the parser for this string. I chose to go with the simpler
straightforward approach to match the behavior of the backend.

As of this commit all non-skipped tests now pass with the lib-parser.