tester: lib-parser now reports filename
I had used a "?" to indicate the filename for locations that the lib-parser
would report in Parser.Compat.hs. This "?" was then output by the tester. The
discrepency is down to megaparsec v. parsec. Megaparsec keeps thi filename in
its SourcePos data type. That datatype is then used by the tester. Parsec does
not do this so as a workaround we have to plumb through the filename to
'locToSrcLoc'.