parser: handle type macro correctly
Lennart pointed out that at the repl if you did:
```
> type{0}
fun(_-9223372036854775808 := 0)<c><succeeds>{_-9223372036854775808}

Why? Because the rewrite pass was creating a fresh variable for the type
macro. Which is no longer needed in the library version of this parser.

This commit removes that fresh variable and just allows the FrontEnd desugarer
to deal with the macro.