parser: correct unexpected `ascii-code` output
- follow up to #105

Previously the parser would issue errors such as "unexpected 41" because Parsec
wraps all its errors in a String and builds in a Show instance. This means that
our bytestring streams are interpreted and output in the error case as their
ascii codes rather than the human friendly versions. This commit fixes up these
errors to correctly output the human friendly version:  "unexpected ')'".