parser: correct parsing mismatch between tester and repl
- closes #103

This expression `((1,2) ||| (3,4))` did not correctly parse in the tester but
did correctly parse in the repl. Why? because the repl cleaned whitespace while
the tester did not. Specifically this simplified expression ` (1,2)[0]` would
also fail with error:
```
"versetests/densempom.versetest" (line 111, column 47):
unexpected "["
expecting "<#" or "<#>"
```
even though the error is that the leading space was not properly consumed.

This commit fixes that by making the testds parser in tester call `pcBraces`
which consumes whitespace after the leading lbracket. Additionally, I added
tests to test that whitespace is consumed between applications.

Lastly, I added some utilities to the parser test suite that allow more testing
of parsing directly to FrontEnd.Expr.